Cast attn_output to h dtype in paged_llama_attention_block (#1761)
Prevent h to be casted to fp32 in the attention block.

Performance on llama prefill 8b fp8, acc f16

Before change:
```
-------------------------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------------
BM_prefill_bs4/process_time/real_time               239 ms          239 ms            3 items_per_second=4.18696/s
BM_prefill_bs4/process_time/real_time               240 ms          293 ms            3 items_per_second=4.16955/s
BM_prefill_bs4/process_time/real_time               240 ms          240 ms            3 items_per_second=4.16825/s
BM_prefill_bs4/process_time/real_time_mean          240 ms          258 ms            3 items_per_second=4.17492/s
BM_prefill_bs4/process_time/real_time_median        240 ms          240 ms            3 items_per_second=4.16955/s
BM_prefill_bs4/process_time/real_time_stddev      0.598 ms         30.8 ms            3 items_per_second=0.0104445/s
BM_prefill_bs4/process_time/real_time_cv           0.25 %         11.98 %             3 items_per_second=0.25%
```
After change:
```
-------------------------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------------------
BM_prefill_bs4/process_time/real_time               230 ms          230 ms            3 items_per_second=4.35173/s
BM_prefill_bs4/process_time/real_time               231 ms          281 ms            3 items_per_second=4.33825/s
BM_prefill_bs4/process_time/real_time               230 ms          230 ms            3 items_per_second=4.34477/s
BM_prefill_bs4/process_time/real_time_mean          230 ms          247 ms            3 items_per_second=4.34491/s
BM_prefill_bs4/process_time/real_time_median        230 ms          230 ms            3 items_per_second=4.34477/s
BM_prefill_bs4/process_time/real_time_stddev      0.357 ms         29.4 ms            3 items_per_second=6.74204m/s
BM_prefill_bs4/process_time/real_time_cv           0.16 %         11.89 %             3 items_per_second=0.16%
```

+++ sharktank/sharktank/layers/paged_llama_attention_block.py
@@ -318,5 +318,5 @@ class PagedLlamaAttentionBlock(ThetaLayer):
         attn_output = self.attn_output(attn_output)
         attn_output = self.attn_output_norm(attn_output)
 
-        h = h + attn_output
+        h = h + attn_output.to(dtype=h.dtype)
         return h
